splunk/requirements.txt
flask
pyyaml
pyopenssl 
requests
beautifulsoup4